WIP redux async test fail
