+++ package.json
@@ -34,7 +34,9 @@
     "not op_mini all"
   ],
   "devDependencies": {
+    "fetch-mock": "^7.0.7",
     "react-test-renderer": "^16.5.2",
-    "redux-devtools-extension": "^2.13.5"
+    "redux-devtools-extension": "^2.13.5",
+    "redux-mock-store": "^1.5.3"
   }
 }
