+++ src/__tests__/HomeScreenContainer.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import HomeScreenContainer from '../containers/HomeScreenContainer';
 
-it('HomescreenContainer renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<HomeScreenContainer />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
+// it('HomescreenContainer renders without crashing', () => {
+//   const div = document.createElement('div');
+//   ReactDOM.render(<HomeScreenContainer />, div);
+//   ReactDOM.unmountComponentAtNode(div);
+// });
